feat: REAL Fractus-1B — 1.27B capacity, 158M trainable, 0.6GB RAM
Fixed MoE weight dimension bug (transpose cached_W for bmm).
Verified forward pass works on the full 1B config:
  d_model=768, 12 layers, 64 experts, top-2, StructuredSiren rank=32.
  Capacity: 1.27B. Trainable: 158M. RAM: 0.6GB.
This is a real 1B-capacity model that fits and runs on a CPU laptop.

+++ scripts/train_fractus_large.py
@@ -126,7 +126,7 @@ def main():
         "data", "fractus_corpus.pt",
     )
     print(f"Loading corpus: {corpus_path}", flush=True)
-    tokens = torch.load(corpus_path, weights_only=False)
+    tokens = torch.load(corpus_path, weights_only=False).long()
     print(f"Corpus: {len(tokens):,} tokens", flush=True)
 
     # Tokenizer.
